Unbound variable bug fix
This commit fixes a bug where if the user tries to
evaluate an unbound variable and then does a .show_vars,
the program crashes. The problem was that the tokenizer
inserts the variable into the var table regardless of
whether it is bound. We never remove the variable.
This table maps the variable name to a handle used to identify
the variable. When we go to print the variable's value,
it was assumed that all tokens in the var table would have
corresponding values in the var_map. This assumption was incorrect
and led to a crash when we did a blind unwrap of the option
from the var_map. The fix is to simply check for
the variable not having a value in the var_map.